v107.18
* fix EXLOROM detection
* fix new CPU masking error

+++ bsnes/processor/wdc65816/wdc65816.hpp
@@ -10,8 +10,8 @@ struct WDC65816 {
   virtual auto idle() -> void = 0;
   virtual auto idleBranch() -> void {}
   virtual auto idleJump() -> void {}
-  virtual auto read(uint addr) -> uint8 = 0;
-  virtual auto write(uint addr, uint8 data) -> void = 0;
+  virtual auto read(uint24 addr) -> uint8 = 0;
+  virtual auto write(uint24 addr, uint8 data) -> void = 0;
   virtual auto lastCycle() -> void = 0;
   virtual auto interruptPending() const -> bool = 0;
   virtual auto interrupt() -> void;
